requirements.txt
# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Terminal UI
rich>=13.7.0
textual>=0.47.0
prompt-toolkit>=3.0.43

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Data handling
numpy>=1.24.0
pandas>=2.1.0
pillow>=10.0.0
h5py>=3.10.0

# Model utilities
einops>=0.7.0
transformers>=4.36.0
tokenizers>=0.15.0

# Monitoring and logging
tensorboard>=2.15.0
wandb>=0.16.0
tqdm>=4.66.0

# Performance
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
psutil>=5.9.0
GPUtil>=1.4.0

# Development
pytest>=7.4.0
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
